// src/utils/api/profile/types/types.ts
export interface GenderType {
  id: number;
  name: string;
}

export interface ICountryType {
  id: number;
  code: string;
  flag: string;
  name: string;
  dialingCode: string;
}
export interface LanguageType {
  id: number;
  name: string;
}
export interface LivelsOptionType {
  id: 1;
  name: string;
  description: string;
}

export interface CityType {
  id: number;
  name: string;
  description: string;
}

export interface RoleType {
  id: number;
  name: string;
  description: string;
}

export interface MissionInstitutionType {
  id: number;
  name: string;
  description: string;
}

export interface yearsOfExperiencesType {
  id: number;
  name: string;
  description: string;
}